DAOS-14348 client: remove dcache entry when a dir is deleted
Integrate reviewers comments:
- Put dcache code into a seperate file
- Add field's documentation

Required-githooks: true
Features: pil4dfs
